website/requirements.txt
python-dotenv
python-json-logger
requests
flask
flask-login
sqlalchemy
flask-cors
gunicorn